+++ lib/widgets/dialog_action.dart
@@ -1,0 +1,45 @@
+/*
+ *******************************************************************************
+ Package:  biorhythmmm
+ Class:    dialog_action.dart
+ Author:   Nathan Cosgray | https://www.nathanatos.com
+ -------------------------------------------------------------------------------
+ Copyright (c) 2024 Nathan Cosgray. All rights reserved.
+
+ This source code is licensed under the BSD-style license found in LICENSE.txt.
+ *******************************************************************************
+*/
+
+// Biorhythmmm
+// - Platform aware dialog action
+
+import 'dart:io' show Platform;
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+// Dialog action button appropriate to platform
+Widget adaptiveDialogAction({
+  bool isDefaultAction = false,
+  bool isDestructiveAction = false,
+  required String text,
+  required Function()? onPressed,
+}) {
+  if (Platform.isIOS) {
+    return CupertinoDialogAction(
+      isDefaultAction: isDefaultAction,
+      isDestructiveAction: isDestructiveAction,
+      onPressed: onPressed,
+      child: Text(text),
+    );
+  } else {
+    return isDestructiveAction
+        ? TextButton(
+            onPressed: onPressed,
+            child: Text(text),
+          )
+        : FilledButton.tonal(
+            onPressed: onPressed,
+            child: Text(text),
+          );
+  }
+}
